schematic sync and new PCB layout and footprint
update the project and sync the schematic to the pcb with new lib for the AVIP connector.
AVIP was done before in pads and vias fixed it up properly with a valid footprint

Replaced the transistors with 2n2222 could not source the original ones and have been discontinued (note new layout for transistors to suit.)

Fixed up a few DRC rules and made solder mask cover vias.